[Bronze IV] Title: 할인이 필요해, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/34323. 할인이 필요해/할인이 필요해.cc
@@ -1,9 +1,8 @@
 #include <iostream>
-#include <cmath>
 using namespace std;
 int main(void) {
     long n,m,s;
     cin>>n>>m>>s;
-    cout<<min(s*m,(long)(s*(m+1)*(100-n)/100.0));
+    cout<<min(s*m,s*(m+1)*(100-n)/100);
     return 0;
 }
